Set up controller, routes, and view to handle OAuth with Twitter
Specifically, added create and destroy actions in the sessions
controller; create just dumps out the contents of the OAuth
hash returned by Twitter. Destroy will eventually destroy the
session for a logged in user.

The routes file has been modified to handle any request callback
from any Omniauth provider strategy.

Finally, there is a link to log in on the application layout.
Eventually that should be dynamic based on whether someone is
logged in or not.